+++ extensions/opera/src/components/index.ts
@@ -1,13 +1,3 @@
-import { HistoryItemActions, NewTabActions, TabListItemActions } from "./actions";
-import { HistoryItem, TabListItem } from "./list";
-
-export class OperaActions {
-  public static NewTab = NewTabActions;
-  public static TabList = TabListItemActions;
-  public static TabHistory = HistoryItemActions;
-}
-
-export class OperaListsItems {
-  public static TabList = TabListItem;
-  public static TabHistory = HistoryItem;
-}
+export * from "./OperaActions";
+export * from "./OperaListItems";
+export * from "./NotInstalled";