+++ extensions/opera/src/constants.ts
@@ -1,0 +1,14 @@
+export const defaultOperaProfilePath = ["Application Support", "com.operasoftware.Opera"];
+
+export const DownloadOperaText = `
+  # 🚨Error: Opera browser is not installed
+  ## This extension depends on Opera browser. You must install it to continue.
+  
+  [Click here](https://opera.com/download/) to download manually.
+  
+  [![Opera](https://www-static-sites.operacdn.com/wp-content/uploads/sites/6/2022/02/Opera-Desktop-Cover-1.png)]()
+`;
+
+export const DEFAULT_ERROR_TITLE = "An Error Occurred";
+
+export const OPERA_NOT_INSTALLED_MESSAGE = "Opera not installed";