+++ extensions/opera/CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add Opera Open Tab action
 - Add Opera Search Tabs action
 - Add Opera Search History action
+- Add Opera Search Bookmarks action